edge/requirements.txt
python-dotenv
python-scapy